create projectCreate page

+++ database/src/main/java/entity/Project.java
@@ -8,16 +8,16 @@ public class Project {
 	private Long id;
 	private String name;
 	private String description;
-	private Integer adminId;
+	private Long adminId;
 	LocalDate deadline;
-	public Project(Long id, String name, String description, Integer adminId, LocalDate deadline) {
+	public Project(Long id, String name, String description, Long adminId, LocalDate deadline) {
         this.id = id;
         this.name = name;
         this.description = description;
         this.adminId = adminId;
         this.deadline = deadline;
     }
-	public Project( String name, String description, Integer adminId, LocalDate deadline) {
+	public Project( String name, String description, Long adminId, LocalDate deadline) {
         this.name = name;
         this.description = description;
         this.adminId = adminId;
@@ -36,10 +36,10 @@ public class Project {
 	public void setName(String name) {
 		this.name = name;
 	}
-	public Integer getAdminId() {
+	public Long getAdminId() {
 		return adminId;
 	}
-	public void setAdminId(Integer adminId) {
+	public void setAdminId(Long adminId) {
 		this.adminId = adminId;
 	}
 	public String getDescription() {
